student_visit/requirements.txt
django
djangorestframework
face_recognition